changed name file to climbing-crag
